Rename: config 파일들을 global config 패키지 대신 jooq, jpa, swagger 패키지 사용
